+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
 - cosmwasm-std: New const methods `Uint128::to_be_bytes`/`::to_le_bytes`.
 - cosmwasm-std: New const conversion methods `Uint256::from_uint128` and
   `Uint512::from_uint256`.
+- cosmwasm-std: New getters `Decimal{,256}::atomics()` and
+  `Decimal{,256}::decimal_places()`.
 
 ### Changed